Hermes console 2 | ACE editor for debugging subscription filters - config moved to component folder.
